RockPaperScissors Game
# Rock-Paper-Scissors Game

Build a game in which two players compete in a game of Rock-Paper-Scissors with different
strategies. Who will win more rounds?

## Rules

- Scissors beats Paper
- Rock beat Scissors
- Paper beats Rock
- If both players choose the same, the round is counted as a tie.

Implement two players:

 - Player A always chooses Paper
 - Player B chooses randomly

The game consists of 100 rounds of above two players competing. The output of the program
should be like the following:

"Player A wins 31 of 100 games"
"Player B wins 37 of 100 games"
"Tie: 32 of 100 games"